+++ src/avt/Expressions/Math/avtArctanExpression.C
@@ -113,7 +113,7 @@ avtArctanExpression::DoOperation(vtkDataArray *in, vtkDataArray *out,
     {
         for (int j = 0 ; j < ncomponents ; j++)
         {
-            float val = in->GetComponent(i, j);
+            double val = in->GetComponent(i, j);
             out->SetComponent(i, j, atan(val));
         }
     }
